Separate eval runner configuration
